requirements.txt
openai
langchain
langchain-core
langchain_community
langchain_groq
python-dotenv
groq
pydantic-settings
supabase
fastapi
uvicorn
python-multipart